change type from decimal to float becouse of benchmarks

+++ db/migrate/20180806123120_create_expense_stuffs.rb
@@ -2,7 +2,7 @@ class CreateExpenseStuffs < ActiveRecord::Migration[5.1]
   def change
     create_table :expense_stuffs do |t|
       t.string :name
-      t.money :cost
+      t.float :cost
       t.datetime :paydeadline
       t.timestamps
     end
