+++ db/migrate/20180806135739_add_amount_types_to_expense_stuff.rb
@@ -1,11 +1,11 @@
 class AddAmountTypesToExpenseStuff < ActiveRecord::Migration[5.1]
   def change
-    add_column :expense_stuffs, :max_cost, :money
-    add_column :expense_stuffs, :min_cost, :money
-    add_column :expense_stuffs, :mid_cost, :money
-    add_column :expense_stuffs, :max_amount, :integer
-    add_column :expense_stuffs, :mid_amount, :integer
-    add_column :expense_stuffs, :min_amount, :integer
-    add_column :expense_stuffs, :amount, :integer
+    add_column :expense_stuffs, :max_cost, :float
+    add_column :expense_stuffs, :min_cost, :float
+    add_column :expense_stuffs, :mid_cost, :float
+    add_column :expense_stuffs, :max_amount, :float
+    add_column :expense_stuffs, :mid_amount, :float
+    add_column :expense_stuffs, :min_amount, :float
+    add_column :expense_stuffs, :amount, :float
   end
 end
